Fix cargo test --workspace compilation errors
11 issues fixed across radium-core and radium-orchestrator:

- dev_loop_integration_test: import WorkspaceRootProvider from file_tools
  (git_extended_tools uses file_tools trait, not terminal_tool trait);
  implement both traits on TestWorkspaceRoot; add explicit Arc type annotations
- session_api_test: gate with #![cfg(feature = "server")] since
  RadiumService is only available under that feature
- multi_agent_integration_test: cast usize arithmetic to u64 for delay_ms field
- workflow_failure_test: use ref binding to avoid partial move of failure_type
- workflow_crud_test: gate on both workflow AND server features
- agent_delegation_e2e_test: update to new AgentConfig fields and
  TriggerBehaviorConfig → AgentTriggerBehavior API
- oversight_integration_test: add metadata/tool_calls fields to ModelResponse
- sandbox_integration_test: import SandboxManager trait into scope
- agent_selector.rs tests: update to new orchestrator API (AgentOutput,
  ModelError, register_agent, select_agent return type)
- metrics_aggregator.rs tests: update RoutingFeedbackRecord fields
  (timestamp: u64, add routing_method and comment)

+++ crates/radium-orchestrator/tests/multi_agent_integration_test.rs
@@ -335,7 +335,7 @@ async fn test_concurrent_execution_result_ordering() {
     let tools: Vec<Tool> = (0..5)
         .map(|i| {
             // Reverse delay order: tool 0 takes longest, tool 4 takes shortest
-            let delay_ms = (5 - i) * 20;
+            let delay_ms = ((5 - i) * 20) as u64;
             Tool::new(
                 &format!("tool_{}", i),
                 &format!("tool_{}", i),
